refactoring db_utils
