+++ backend/sample.test.env
@@ -1,4 +1,4 @@
-DB_HOST='localhost'
+DB_HOST='unstract-db'
 DB_USER='unstract_dev'
 DB_PASSWORD='unstract_pass'
 DB_NAME='unstract_db'
